django-backend/src/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-debug-toolbar==4.2.0
django-filter==23.3
django-resized==1.0.2
djangorestframework==3.14.0
gunicorn==21.2.0
Markdown==3.5
packaging==23.2
Pillow==10.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
